requirements.txt
streamlit 
yfinance 
plotly 
pandas 
google-generativeai